:art: Moves new block to match other core
