+++ lib/core/lg-new-block.js
