+++ lib/lost-align.js
@@ -1,4 +1,4 @@
-var newBlock = require('./new-block.js');
+var newBlock = require('./core/lg-new-block.js');
 
 module.exports = function lostAlign(css, settings) {
   css.walkDecls('lost-align', function alignDirectionFunction(decl) {
