+++ lib/lost-column.js
@@ -1,4 +1,4 @@
-var newBlock = require('./new-block.js');
+var newBlock = require('./core/lg-new-block.js');
 
 var lgLogic = require('./core/lg-logic');
 var lgUtils = require('./core/lg-utilities');
