+++ lib/lost-masonry-wrap.js
@@ -1,4 +1,4 @@
-var newBlock = require('./new-block.js');
+var newBlock = require('./core/lg-new-block.js');
 
 module.exports = function lostMasonryWrapDecl(css, settings) {
   css.walkDecls('lost-masonry-wrap', function lostMasonryWrapDeclFunction(
