+++ lib/lost-utility.js
@@ -1,4 +1,4 @@
-var newBlock = require('./new-block.js');
+var newBlock = require('./core/lg-new-block.js');
 var getColorValue = require('./core/lg-utilities').getColorValue;
 
 function unitsMatch() {
